+++ roguelike_project/engine/game/systems/map/geometry.py
