+++ roguelike_project/engine/game/systems/map/tile_loader.py
@@ -4,7 +4,7 @@ from roguelike_project.utils.loader import load_image
 from roguelike_project.config import DEBUG
 import random
 from roguelike_project.config import TILE_SIZE
-from roguelike_project.map.tile import Tile
+from roguelike_project.engine.game.systems.map.tile import Tile
 
 def load_tile_images(theme="default"):
     path = "assets/tiles"
